Changed end point to CreateEvent
