+++ Travlendar/frontend/src/Routes.js
@@ -1,7 +1,7 @@
 export const HOMEPAGE = '/';
 export const SIGNUP = '/Signup';
 export const LOGIN = '/Login';
-export const MAPCONTAINER = '/MapContainer';
+export const CREATE_EVENT = '/CreateEvent';
 export const EVENTS_LIST = '/listevent';
 export const EVENTS_MAP= '/map';
 export const EMAIL = '/emailalert';
